[#6027, #6190] Build standard document creation dialog in AppV2
Creates a standard `CreateDocumentDialog` that replaces core's
`Document#createDialog` to give a shared implementation using
the system's design across actors, items, activities, and
advancement. This allows items and actors to take advantage of
the `_createDialogTypes` and `_createDialogData` methods that are
used by activites and advancement to provide icons and hints to
options.

Closes #6027
Closes #6190

 ### Todo
- [ ] Provide type-specific icons for each actor type
- [ ] Provide hints for item types
- [ ] Expand dialog to `JournalEntryPage` & `RegionBehavior`